[frameworks][av] fix -Wimplicit-int-float-conversion
IEEE754 single precision cannot precisely represent
INT32_MAX/UINT_MAX/RAND_MAX.

IEEE754 double precision cannot precisely represent INT64_MAX.

As such, implicit conversions from operations with floating point and
integral operands may result in a loss of precision.

Accept the imprecision with by making the existing implicit casts
explicit.

Bug: 139945549
Test: mm

+++ media/libstagefright/MediaClock.cpp
@@ -281,7 +281,7 @@ void MediaClock::processTimers_l() {
             it = mTimers.erase(it);
         } else {
             if (mPlaybackRate != 0.0
-                && (double)diffMediaUs < INT64_MAX * (double)mPlaybackRate) {
+                && (double)diffMediaUs < (double)INT64_MAX * (double)mPlaybackRate) {
                 int64_t targetRealUs = diffMediaUs / (double)mPlaybackRate;
                 if (targetRealUs < nextLapseRealUs) {
                     nextLapseRealUs = targetRealUs;
